requirements.txt
arcade==2.6.17
numpy>=1.26
pillow~=9.3.0
dataclasses-json>=0.6
typer>=0.9
pyyaml>=6.0
rich>=13.7
noise